Change temperature and humidity to float fields

+++ api/models.py
@@ -25,12 +25,10 @@ class Telemetry(TimestampModel):
     oxygen = models.IntegerField()
     ion_radiation = models.IntegerField(verbose_name='Ionizing radiation')
 
-    humidity = models.DecimalField(
-        max_digits=5, decimal_places=2,
+    humidity = models.FloatField(
         help_text='0 to 100%, resolution 0.04%')
 
-    temperature = models.DecimalField(
-        max_digits=5, decimal_places=2,
+    temperature = models.FloatField(
         help_text='-40 to 125℃, resolution 0.01℃')
 
     pressure = models.FloatField(help_text='Air pressure [hPa]')
